make directive as a plugin
