+++ src/App.vue
@@ -11,8 +11,6 @@
 </template>
 
 <script>
-
-
 export default {
   name: 'app',
   data () {
